[*] Refactored pathways;

+++ app/src/main/java/legalFreedom/java/model/data/dto/Book.java
@@ -17,7 +17,7 @@ public class Book {
         this.bookId = bookId;
     }
 
-    public float getId() {
+    public long getId() {
         return id;
     }
